src/requirements.txt
fastapi[all]
gunicorn
uvicorn
httpx
python-dotenv
pydantic-settings
pytest-cov
pytest
black
flake8